Create IncomeExpenseListItem and reorganize the structure for report page
